# frontend-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContentGuard API - Content Protection Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContentType = (typeof ContentType)[keyof typeof ContentType];

export const ContentType = {
  paper: "paper",
  course: "course",
  image: "image",
  document: "document",
  video: "video",
} as const;

export type ContentStatus = (typeof ContentStatus)[keyof typeof ContentStatus];

export const ContentStatus = {
  active: "active",
  monitoring: "monitoring",
  archived: "archived",
} as const;

export interface Content {
  id: string;
  title: string;
  type: ContentType;
  description?: string;
  contentHash: string;
  fileSize?: number;
  author: string;
  organization: string;
  registeredAt: string;
  blockchainTxHash?: string | null;
  ipfsHash?: string | null;
  detectionCount: number;
  status: ContentStatus;
  similarityThreshold: number;
}

export type RegisterContentRequestType =
  (typeof RegisterContentRequestType)[keyof typeof RegisterContentRequestType];

export const RegisterContentRequestType = {
  paper: "paper",
  course: "course",
  image: "image",
  document: "document",
  video: "video",
} as const;

export interface RegisterContentRequest {
  title: string;
  type: RegisterContentRequestType;
  description?: string;
  author: string;
  organization: string;
  contentHash: string;
  similarityThreshold?: number;
  registerOnBlockchain?: boolean;
}

export interface ContentListResponse {
  items: Content[];
  total: number;
  page: number;
  limit: number;
}

export type DetectionDetectionType =
  (typeof DetectionDetectionType)[keyof typeof DetectionDetectionType];

export const DetectionDetectionType = {
  exact: "exact",
  paraphrase: "paraphrase",
  partial: "partial",
  visual: "visual",
} as const;

export type DetectionStatus =
  (typeof DetectionStatus)[keyof typeof DetectionStatus];

export const DetectionStatus = {
  pending: "pending",
  confirmed: "confirmed",
  dismissed: "dismissed",
} as const;

export interface Detection {
  id: string;
  contentId: string;
  contentTitle: string;
  similarityScore: number;
  detectionType: DetectionDetectionType;
  sourceUrl: string;
  sourcePlatform: string;
  detectedAt: string;
  status: DetectionStatus;
  excerpt: string;
  aiAnalysis?: string | null;
}

export type UpdateDetectionStatusRequestStatus =
  (typeof UpdateDetectionStatusRequestStatus)[keyof typeof UpdateDetectionStatusRequestStatus];

export const UpdateDetectionStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  dismissed: "dismissed",
} as const;

export interface UpdateDetectionStatusRequest {
  status: UpdateDetectionStatusRequestStatus;
}

export interface DetectionListResponse {
  items: Detection[];
  total: number;
  page: number;
  limit: number;
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  new_detection: "new_detection",
  high_similarity: "high_similarity",
  blockchain_registered: "blockchain_registered",
  weekly_summary: "weekly_summary",
} as const;

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Alert {
  id: string;
  type: AlertType;
  title: string;
  message: string;
  contentId?: string | null;
  detectionId?: string | null;
  read: boolean;
  createdAt: string;
  severity: AlertSeverity;
}

export interface AlertListResponse {
  items: Alert[];
  total: number;
  unreadCount: number;
  page: number;
  limit: number;
}

export interface AffectedContentStat {
  contentId: string;
  title: string;
  detectionCount: number;
  avgSimilarity: number;
}

export interface PlatformStat {
  platform: string;
  count: number;
  percentage: number;
}

export interface AnalyticsOverview {
  totalContent: number;
  activeMonitoring: number;
  totalDetections: number;
  confirmedInfringements: number;
  pendingReview: number;
  avgSimilarityScore: number;
  blockchainRegistered: number;
  mostAffectedContent: AffectedContentStat[];
  platformBreakdown: PlatformStat[];
}

export type PropagationNodeType =
  (typeof PropagationNodeType)[keyof typeof PropagationNodeType];

export const PropagationNodeType = {
  origin: "origin",
  copy: "copy",
  partial: "partial",
} as const;

export interface PropagationNode {
  id: string;
  label: string;
  type: PropagationNodeType;
  platform: string;
  url: string;
  similarityScore: number;
  detectedAt: string;
}

export interface PropagationLink {
  source: string;
  target: string;
  weight: number;
}

export interface PropagationData {
  nodes: PropagationNode[];
  links: PropagationLink[];
}

export interface TrendPoint {
  date: string;
  detections: number;
  confirmed: number;
  dismissed: number;
}

export interface TypeTrendStat {
  type: string;
  count: number;
}

export interface TrendsData {
  daily: TrendPoint[];
  byType: TypeTrendStat[];
}

export interface BlockchainRegisterRequest {
  contentId: string;
  walletAddress: string;
}

export interface BlockchainRecord {
  contentId: string;
  txHash: string;
  blockNumber: number;
  network: string;
  ipfsHash: string;
  timestamp: string;
  ownerAddress: string;
}

export interface OwnershipVerification {
  contentId: string;
  verified: boolean;
  ownerAddress?: string | null;
  txHash?: string | null;
  registeredAt?: string | null;
  network?: string | null;
}

export type ListContentParams = {
  page?: number;
  limit?: number;
  type?: ListContentType;
};

export type ListContentType =
  (typeof ListContentType)[keyof typeof ListContentType];

export const ListContentType = {
  paper: "paper",
  course: "course",
  image: "image",
  document: "document",
  video: "video",
} as const;

export type ListDetectionsParams = {
  page?: number;
  limit?: number;
  contentId?: string;
  status?: ListDetectionsStatus;
  minSimilarity?: number;
};

export type ListDetectionsStatus =
  (typeof ListDetectionsStatus)[keyof typeof ListDetectionsStatus];

export const ListDetectionsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  dismissed: "dismissed",
} as const;

export type ListAlertsParams = {
  page?: number;
  limit?: number;
  read?: boolean;
};

export type GetPropagationDataParams = {
  contentId?: string;
};