# frontend-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ContentGuard API - Content Protection Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /web3/register:
    post:
      operationId: registerOnBlockchain
      tags: [web3]
      summary: Register content hash on blockchain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockchainRegisterRequest"
      responses:
        "201":
          description: Registered on blockchain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockchainRecord"
  /web3/verify/{contentId}:
    get:
      operationId: verifyOwnership
      tags: [web3]
      summary: Verify content ownership on blockchain
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnershipVerification"
components:
  schemas:
...
    BlockchainRegisterRequest:
      type: object
      properties:
        contentId:
          type: string
        walletAddress:
          type: string
      required: [contentId, walletAddress]
    BlockchainRecord:
      type: object
      properties:
        contentId:
          type: string
        txHash:
          type: string
        blockNumber:
          type: integer
        network:
          type: string
        ipfsHash:
          type: string
        timestamp:
          type: string
          format: date-time
        ownerAddress:
          type: string
      required: [contentId, txHash, blockNumber, network, ipfsHash, timestamp, ownerAddress]
    OwnershipVerification:
      type: object
      properties:
        contentId:
          type: string
        verified:
          type: boolean
        ownerAddress:
          type: string
          nullable: true
        txHash:
          type: string
          nullable: true
        registeredAt:
          type: string
          nullable: true
        network:
          type: string
          nullable: true
      required: [contentId, verified]